removed male-gendered language
hi, i'm a nonbinary engineer, and having gendered language at all, *especially* "Sir," in a readme like this is incredibly off-putting and reinforces the idea that most programmers are men. gendered language isn't necessary, even if you're just trying to be playful to the reader. this is just a suggestion. thank you!