make us of opts in parse_cron and raise error
`parse_cron` method was expecting an "required" option hash as parameter but it was not used at all. This commit changes the parameter to an "optional" one and actually uses it to determine if the method should return nil or raise an error.(as other parse methods). `Fugit.do_parse_cron` is just like `parse_cron` but it raises an "ArgumentError" if the string cannot be parsed.